refactor: group files for next and past events
