+++ modules/sponsors/Sponsors.module.css
@@ -19,5 +19,4 @@
 
 .sponsors a img {
   max-height: 80px;
-  aspect-ratio: auto;
 }
