+++ pages/index.tsx
@@ -6,14 +6,13 @@ import { dataOverride } from '../data/data-override';
 
 import type { Event } from '../modules/event/types';
 import { LyonJSHead } from '../modules/header/LyonJSHead';
-import { EventCard } from '../modules/event/EventCard';
 import { fetchMeetupEvents } from '../modules/meetup/api';
 import { H2 as Heading2 } from '../modules/atoms/remark/Titles';
 import styles from '../modules/home/Home.module.css';
 import { Hero } from '../modules/home/Hero';
-import { NoNextEvent } from '../modules/event/NoNextEvent';
 import { ButtonLink } from '../modules/atoms/button/Button';
 import { Sponsors } from '../modules/sponsors/Sponsors';
+import { NextEvent } from '../modules/event/next-event/NextEvent';
 
 export const Article: FC<PropsWithChildren> = ({ children }) => (
   <article className={styles.article}>{children}</article>
@@ -26,13 +25,11 @@ export const H2: FC<PropsWithChildren> = ({ children }) => (
 );
 
 const SeePastEvents = () => (
-  <ButtonLink
-    variant="secondary"
-    href={`/evenements-precedents/${new Date().getFullYear()}`}
-    className={styles.seePastEvents}
-  >
-    Voir les événements passés <span aria-hidden="true">&rarr;</span>
-  </ButtonLink>
+  <div className={styles.seePastEvents}>
+    <ButtonLink variant="secondary" href={`/evenements-precedents/${new Date().getFullYear()}`}>
+      Voir les événements passés <span aria-hidden="true">&rarr;</span>
+    </ButtonLink>
+  </div>
 );
 
 type Props = { nextEvent: Event };
@@ -42,14 +39,8 @@ const Home: NextPage<Props> = ({ nextEvent }) => (
     <LyonJSHead />
     <main>
       <Hero />
-
-      <Article>
-        <H2>Prochain évènement</H2>
-        {nextEvent ? <EventCard event={nextEvent} /> : <NoNextEvent />}
-      </Article>
-
+      <NextEvent event={nextEvent} />
       <SeePastEvents />
-
       <Sponsors />
     </main>
   </>
